cleanup analysis-specific kernels
